+++ cli/internal/hasura/commonmetadata/v2commonmetadataops_test.go
@@ -56,7 +56,7 @@ func TestClientCommonMetadataOps_V2ReplaceMetadata(t *testing.T) {
                     "database_url": {
                         "from_env": "HASURA_GRAPHQL_DATABASE_URL"
                     },
-		    "isolation_level": "read-committed",
+		    		"isolation_level": "read-committed",
                     "pool_settings": {
                         "retries": 1,
                         "idle_timeout": 180,