+++ console/src/components/Services/Data/TreeView.tsx
@@ -140,6 +140,7 @@ const SchemaItemsView: React.FC<SchemaItemsViewProps> = ({
   setActiveSchema,
   pathname,
   databaseLoading,
+  ...props
 }) => {
   const [isOpen, setIsOpen] = useState(false);
   const showActiveStyle =
@@ -161,6 +162,7 @@ const SchemaItemsView: React.FC<SchemaItemsViewProps> = ({
         role="button"
         className={styles.padd_bottom_small}
         style={showActiveStyle ? activeStyle : {}}
+        {...props}
       >
         <span
           className={
@@ -262,6 +264,7 @@ const DatabaseItemsView: React.FC<DatabaseItemsViewProps> = ({
                 setActiveSchema={handleSelectSchema}
                 key={key}
                 pathname={pathname}
+                data-test={`${child.name}_schema`}
                 databaseLoading={databaseLoading}
               />
             </li>