+++ console/src/dataSources/types.ts
@@ -159,7 +159,24 @@ export interface Table extends BaseTable {
     table_schema: string;
     definition: RemoteRelationshipDef;
   }[];
+  unique_constraints:
+    | {
+        table_name: string;
+        table_schema: string;
+        constraint_name: string;
+        columns: string[];
+      }[]
+    | null;
 }
+
+export type Partition = {
+  parent_schema: string;
+  partition_schema: string;
+  partition_name: string;
+  parent_table: string;
+  partition_def: string;
+  partition_key: string;
+};
 
 export type ColumnAction = 'add' | 'modify';
 
@@ -218,6 +235,23 @@ export type SupportedFeaturesType = {
     };
     modify: {
       enabled: boolean;
+      readOnly?: boolean;
+      comments?: boolean;
+      columns?: boolean;
+      columns_edit?: boolean;
+      computedFields?: boolean;
+      primaryKeys?: boolean;
+      primaryKeys_edit?: boolean;
+      foreginKeys?: boolean;
+      foreginKeys_edit?: boolean;
+      uniqueKeys?: boolean;
+      uniqueKeys_edit?: boolean;
+      triggers?: boolean;
+      checkConstraints?: boolean;
+      customGqlRoot?: boolean;
+      setAsEnum?: boolean;
+      untrack?: boolean;
+      delete?: boolean;
     };
     relationships: {
       enabled: boolean;