+++ server/src-lib/Hasura/Backends/BigQuery/Instances/Execute.hs
@@ -3,45 +3,47 @@
 module Hasura.Backends.BigQuery.Instances.Execute () where
 
 import qualified Data.Aeson                                  as Aeson
-import qualified Data.HashMap.Strict.InsOrd                  as OMap
-import qualified Hasura.Backends.BigQuery.DataLoader.Execute as DataLoader
-import qualified Hasura.Backends.BigQuery.DataLoader.Plan    as DataLoader
-import           Hasura.EncJSON
-import           Hasura.Prelude
-import qualified Hasura.RQL.Types.Error                      as RQL
-import qualified Hasura.SQL.AnyBackend                       as AB
-
 import qualified Data.Environment                            as Env
+import qualified Data.HashMap.Strict.InsOrd                  as OMap
+import qualified Data.Text                                   as T
 import qualified Language.GraphQL.Draft.Syntax               as G
 import qualified Network.HTTP.Client                         as HTTP
 import qualified Network.HTTP.Types                          as HTTP
 
+import qualified Hasura.Backends.BigQuery.DataLoader.Execute as DataLoader
+import qualified Hasura.Backends.BigQuery.DataLoader.Plan    as DataLoader
+import qualified Hasura.RQL.Types.Error                      as RQL
+import qualified Hasura.SQL.AnyBackend                       as AB
 import qualified Hasura.Tracing                              as Tracing
 
 import           Hasura.Backends.BigQuery.Plan
+import           Hasura.EncJSON
 import           Hasura.GraphQL.Context
 import           Hasura.GraphQL.Execute.Backend
 import           Hasura.GraphQL.Parser
+import           Hasura.Prelude
 import           Hasura.RQL.Types
 import           Hasura.Session
 
--- MultiplexedQuery
+
 instance BackendExecute 'BigQuery where
-  type PreparedQuery  'BigQuery = Text
-  type ExecutionMonad 'BigQuery = Tracing.TraceT (ExceptT QErr IO)
+  type PreparedQuery    'BigQuery = Text
   type MultiplexedQuery 'BigQuery = Void
+  type ExecutionMonad   'BigQuery = Tracing.TraceT (ExceptT QErr IO)
   getRemoteJoins = const []
 
-  mkDBQueryPlan = msDBQueryPlan
-  mkDBMutationPlan = msDBMutationPlan
+  mkDBQueryPlan = bqDBQueryPlan
+  mkDBMutationPlan = bqDBMutationPlan
   mkDBSubscriptionPlan _ _ _ _ =
     throwError $ RQL.internalError "Cannot currently perform subscriptions on BigQuery sources."
-  mkDBQueryExplain _ _ _ _ _ = throwError $ RQL.internalError "Cannot currently retrieve query execution plans on BigQuery sources."
-  mkLiveQueryExplain _ = throwError $ RQL.internalError "Cannot currently retrieve query execution plans on BigQuery sources."
+  mkDBQueryExplain = bqDBQueryExplain
+  mkLiveQueryExplain _ =
+    throwError $ RQL.internalError "Cannot currently retrieve query execution plans on BigQuery sources."
+
 
 -- query
 
-msDBQueryPlan
+bqDBQueryPlan
   :: forall m.
      ( MonadError QErr m
      )
@@ -54,7 +56,7 @@ msDBQueryPlan
   -> SourceConfig 'BigQuery
   -> QueryDB 'BigQuery (UnpreparedValue 'BigQuery)
   -> m ExecutionStep
-msDBQueryPlan _env _manager _reqHeaders userInfo _directives sourceName sourceConfig qrf = do
+bqDBQueryPlan _env _manager _reqHeaders userInfo _directives sourceName sourceConfig qrf = do
   select <- planNoPlan userInfo qrf
   let (!headAndTail, !plannedActionsList) =
         DataLoader.runPlan
@@ -103,9 +105,10 @@ recordSetToEncJSON DataLoader.RecordSet {rows} =
         -- a record in it.
         DataLoader.RecordOutputValue !record -> encJFromRecord record
 
+
 -- mutation
 
-msDBMutationPlan
+bqDBMutationPlan
   :: forall m.
      ( MonadError QErr m
      )
@@ -118,5 +121,28 @@ msDBMutationPlan
   -> SourceConfig 'BigQuery
   -> MutationDB 'BigQuery (UnpreparedValue 'BigQuery)
   -> m ExecutionStep
-msDBMutationPlan _env _manager _reqHeaders _userInfo _stringifyNum _sourceName _sourceConfig _mrf =
+bqDBMutationPlan _env _manager _reqHeaders _userInfo _stringifyNum _sourceName _sourceConfig _mrf =
   throw500 "mutations are not supported in BigQuery; this should be unreachable"
+
+
+-- explain
+
+bqDBQueryExplain
+  :: MonadError QErr m
+  => G.Name
+  -> UserInfo
+  -> SourceName
+  -> SourceConfig 'BigQuery
+  -> QueryDB 'BigQuery (UnpreparedValue 'BigQuery)
+  -> m (AB.AnyBackend DBStepInfo)
+bqDBQueryExplain fieldName userInfo sourceName sourceConfig qrf = do
+  actionsForest <- planToForest userInfo qrf
+  pure
+    $ AB.mkAnyBackend
+    $ DBStepInfo @'BigQuery sourceName sourceConfig Nothing
+    $ pure
+    $ encJFromJValue
+    $ ExplainPlan
+        fieldName
+        (Just ("--\n" <> DataLoader.drawActionsForestSQL actionsForest))
+        (Just ("": T.lines (DataLoader.drawActionsForest actionsForest)))