+++ server/src-lib/Hasura/GraphQL/Explain.hs
@@ -5,31 +5,28 @@ module Hasura.GraphQL.Explain
 
 import           Hasura.Prelude
 
-import qualified Data.Aeson                               as J
-import qualified Data.Aeson.TH                            as J
-import qualified Data.HashMap.Strict                      as Map
-import qualified Data.HashMap.Strict.InsOrd               as OMap
-import qualified Language.GraphQL.Draft.Syntax            as G
+import qualified Data.Aeson                             as J
+import qualified Data.Aeson.TH                          as J
+import qualified Data.HashMap.Strict                    as Map
+import qualified Data.HashMap.Strict.InsOrd             as OMap
+import qualified Language.GraphQL.Draft.Syntax          as G
 
-import           Control.Monad.Trans.Control              (MonadBaseControl)
-import qualified Data.Text                                as T
+import           Control.Monad.Trans.Control            (MonadBaseControl)
 
-import qualified Hasura.Backends.BigQuery.DataLoader.Plan as BigQuery
-import qualified Hasura.Backends.BigQuery.Plan            as BigQuery
-import qualified Hasura.GraphQL.Execute                   as E
-import qualified Hasura.GraphQL.Execute.Action            as E
-import qualified Hasura.GraphQL.Execute.Inline            as E
-import qualified Hasura.GraphQL.Execute.Query             as E
-import qualified Hasura.GraphQL.Transport.HTTP.Protocol   as GH
-import qualified Hasura.SQL.AnyBackend                    as AB
+import qualified Hasura.GraphQL.Execute                 as E
+import qualified Hasura.GraphQL.Execute.Action          as E
+import qualified Hasura.GraphQL.Execute.Inline          as E
+import qualified Hasura.GraphQL.Execute.Query           as E
+import qualified Hasura.GraphQL.Transport.HTTP.Protocol as GH
+import qualified Hasura.SQL.AnyBackend                  as AB
 
 import           Hasura.EncJSON
 import           Hasura.GraphQL.Context
 import           Hasura.GraphQL.Execute.Backend
-import           Hasura.GraphQL.Execute.Instances         ()
+import           Hasura.GraphQL.Execute.Instances       ()
 import           Hasura.GraphQL.Parser
 import           Hasura.GraphQL.Transport.Backend
-import           Hasura.GraphQL.Transport.Instances       ()
+import           Hasura.GraphQL.Transport.Instances     ()
 import           Hasura.Metadata.Class
 import           Hasura.RQL.Types
 import           Hasura.Session
@@ -55,34 +52,18 @@ explainQueryField
   => UserInfo
   -> G.Name
   -> QueryRootField UnpreparedValue
-  -> m (Maybe EncJSON)
+  -> m EncJSON
 explainQueryField userInfo fieldName rootField = do
   case rootField of
     RFRemote _ -> throw400 InvalidParams "only hasura queries can be explained"
     RFAction _ -> throw400 InvalidParams "query actions cannot be explained"
-    RFRaw _    -> pure $ Just $ encJFromJValue $ ExplainPlan fieldName Nothing Nothing
-    RFDB sourceName exists   -> dispatch [ do
+    RFRaw _    -> pure $ encJFromJValue $ ExplainPlan fieldName Nothing Nothing
+    RFDB sourceName exists   -> do
       step <- AB.dispatchAnyBackend @BackendExecute exists
         \(SourceConfigWith sourceConfig (QDBR db)) ->
            mkDBQueryExplain fieldName userInfo sourceName sourceConfig db
       AB.dispatchAnyBackend @BackendTransport step runDBQueryExplain
-        ,do
-      -- BigQuery case
-      SourceConfigWith _ (QDBR bqQDB) <-
-        hoistMaybe $ AB.unpackAnyBackend exists
-      lift $ do
-        actionsForest <- BigQuery.planToForest userInfo bqQDB
-        pure $ encJFromJValue $
-          ExplainPlan
-            fieldName
-            (Just ("--\n" <> BigQuery.drawActionsForestSQL actionsForest))
-            (Just ("": T.lines (BigQuery.drawActionsForest actionsForest)))]
-  where dispatch [] = pure Nothing
-        dispatch (x:xs) = do
-          mv <- runMaybeT x
-          case mv of
-            Nothing -> dispatch xs
-            Just v  -> pure (Just v)
+
 
 -- NOTE: This function has a 'MonadTrace' constraint in master, but we don't need it
 -- here. We should evaluate if we need it here.
@@ -112,8 +93,8 @@ explainGQLQuery sc (GQLExplain query userVarsRaw maybeIsRelay) = do
       inlinedSelSet <- E.inlineSelectionSet fragments selSet
       (unpreparedQueries, _, _) <-
         E.parseGraphQLQuery graphQLContext varDefs (GH._grVariables query) inlinedSelSet
-      encJFromList . catMaybes
-        <$> for (OMap.toList unpreparedQueries) (uncurry (explainQueryField userInfo))
+      encJFromList <$>
+        for (OMap.toList unpreparedQueries) (uncurry (explainQueryField userInfo))
 
     G.TypedOperationDefinition G.OperationTypeMutation _ _ _ _ ->
       throw400 InvalidParams "only queries can be explained"