+++ server/src-lib/Hasura/RQL/Types/Common.hs
@@ -392,6 +392,17 @@ mkScalarTypeName PG.PGBoolean  = pure boolScalar
 mkScalarTypeName PG.PGFloat    = pure floatScalar
 mkScalarTypeName PG.PGText     = pure stringScalar
 mkScalarTypeName PG.PGVarchar  = pure stringScalar
+mkScalarTypeName (PG.PGCompositeScalar compositeScalarType) =
+  -- When the function argument is a row type argument
+  -- then it's possible that there can be an object type
+  -- with the table name depending upon whether the table
+  -- is tracked or not. As a result, we get a conflict between
+  -- both these types (scalar and object type with same name).
+  -- To avoid this, we suffix the table name with `_scalar`
+  -- and create a new scalar type
+  (<> $$(G.litName "_scalar")) <$> G.mkName compositeScalarType `onNothing` throw400 ValidationFailed
+  ("cannot use SQL type " <> compositeScalarType <<> " in the GraphQL schema because its name is not a "
+  <> "valid GraphQL identifier")
 mkScalarTypeName scalarType = G.mkName (toSQLTxt scalarType) `onNothing` throw400 ValidationFailed
   ("cannot use SQL type " <> scalarType <<> " in the GraphQL schema because its name is not a "
   <> "valid GraphQL identifier")